+++ Assets/_MAIN/Scripts/Components/Tool/Boots.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
 public class Boots : MonoBehaviour {
-	public int manaCost;
+	public float manaCost;
 	public float bootsSpeed;
 }
