+++ Assets/_MAIN/Scripts/Core/Player.cs
@@ -77,14 +77,14 @@ public class Player : MonoBehaviour {
 	// 	}
 	// }
 
-    public int MaxHP{
-        get{return PlayerPrefs.GetInt(Constants.PlayerPrefKey.PLAYER_STATS_MAXHP, maxHP);}
-        set{PlayerPrefs.SetInt(Constants.PlayerPrefKey.PLAYER_STATS_MAXHP, value);}
+    public float MaxHP{
+        get{return PlayerPrefs.GetFloat(Constants.PlayerPrefKey.PLAYER_STATS_MAXHP, maxHP);}
+        set{PlayerPrefs.SetFloat(Constants.PlayerPrefKey.PLAYER_STATS_MAXHP, value);}
     }
 
-    public int MaxMana{
-        get{return PlayerPrefs.GetInt(Constants.PlayerPrefKey.PLAYER_STATS_MAXMANA, maxMana);}
-        set{PlayerPrefs.SetInt(Constants.PlayerPrefKey.PLAYER_STATS_MAXMANA, value);}
+    public float MaxMana{
+        get{return PlayerPrefs.GetFloat(Constants.PlayerPrefKey.PLAYER_STATS_MAXMANA, maxMana);}
+        set{PlayerPrefs.SetFloat(Constants.PlayerPrefKey.PLAYER_STATS_MAXMANA, value);}
     }
 
 	void OnCollisionEnter2D (Collision2D col) {
